# apps/packages/shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/register": {
      "post": {
        "operationId": "UserController_register",
        "summary": "注册新用户",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功，返回用户基本信息（不含密码）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "用户名已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/profile": {
      "get": {
        "operationId": "UserController_profile",
        "summary": "获取当前登录用户资料",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功，返回当前登录用户的完整资料（含创建时间）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未携带或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "authLogin",
        "summary": "账号密码登录，签发双 Token 并建立全局会话",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "登录成功，返回 Token 与全局会话标识"
          },
          "401": {
            "description": "用户名或密码错误"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/validate": {
      "post": {
        "operationId": "authValidate",
        "summary": "校验 Token 有效性（依据全局会话），供 SP 二次免登录判断",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "返回 valid 与用户信息"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/enterprises": {
      "post": {
        "operationId": "createEnterprise",
        "summary": "创建企业（超级管理员）",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnterpriseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "企业创建成功"
          },
          "409": {
            "description": "企业标识已存在"
          }
        },
        "tags": [
          "enterprises"
        ]
      },
      "get": {
        "operationId": "getEnterprises",
        "summary": "企业列表（超级管理员）",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "企业列表"
          }
        },
        "tags": [
          "enterprises"
        ]
      }
    },
    "/api/v1/enterprises/{id}": {
      "get": {
        "operationId": "getEnterpriseById",
        "summary": "企业详情（超级管理员）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "企业详情"
          },
          "404": {
            "description": "企业不存在"
          }
        },
        "tags": [
          "enterprises"
        ]
      },
      "patch": {
        "operationId": "updateEnterprise",
        "summary": "更新企业（超级管理员）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEnterpriseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "404": {
            "description": "企业不存在"
          },
          "409": {
            "description": "企业标识已存在"
          }
        },
        "tags": [
          "enterprises"
        ]
      },
      "delete": {
        "operationId": "deleteEnterprise",
        "summary": "删除企业（超级管理员，软删除为 disabled）",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "404": {
            "description": "企业不存在"
          }
        },
        "tags": [
          "enterprises"
        ]
      }
    },
...
  },
  "info": {
    "title": "SSO 统一认证中心 API",
    "description": "基于 NestJS 的单点登录（SSO）系统接口文档，符合 OAuth 2.0 / OpenID Connect 标准。\n\n**OAuth 2.0 / OIDC 标准流程（所有 OAuth 端点均在根路径 `/oauth/*`，无 API 前缀）：**\n1. SP 302 跳转 `GET /oauth/authorize?response_type=code&client_id=...&redirect_uri=...&scope=...`\n2. 已登录 → 签发 Authorization Code 并 302 回调 SP；未登录 → 跳转登录页\n3. SP 调用 `POST /oauth/token`（grant_type=authorization_code）换取 access_token + id_token + refresh_token\n4. SP 调用 `GET /oauth/userinfo` 获取用户信息\n5. SP 调用 `POST /oauth/introspect` 校验 Token 有效性\n6. SP 调用 `GET /oauth/endsession` 发起单点登出",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "alice",
            "description": "用户名（3-32 位）"
          },
          "password": {
            "type": "string",
            "example": "Str0ng@Pass",
            "description": "密码（至少 6 位）"
          },
          "email": {
            "type": "string",
            "example": "alice@example.com",
            "description": "邮箱"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UserDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户 ID（UUID）",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "alice"
          },
          "email": {
            "type": "string",
            "description": "邮箱（注册时可选，可能为 null）",
            "example": "alice@example.com",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "账号状态：active 正常 / disabled 禁用",
            "example": "active"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "status"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务状态码",
            "example": 201
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "注册成功"
          },
          "data": {
            "description": "注册成功的用户信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDataDto"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间戳（ISO 8601）",
            "example": "2026-08-04T10:00:00.000Z"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务状态码（等同于 HTTP 状态码）",
            "example": 401
          },
          "message": {
            "type": "string",
            "description": "错误描述",
            "example": "未携带或令牌无效"
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间戳（ISO 8601）",
            "example": "2026-08-04T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "请求路径",
            "example": "/api/v1/users/profile"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "UserProfileDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户 ID（UUID）",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "alice"
          },
          "email": {
            "type": "string",
            "description": "邮箱（注册时可选，可能为 null）",
            "example": "alice@example.com",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "账号状态：active 正常 / disabled 禁用",
            "example": "active"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间（ISO 8601）",
            "example": "2026-08-04T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "status",
          "createdAt"
        ]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "success"
          },
          "data": {
            "description": "当前登录用户资料",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserProfileDataDto"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间戳（ISO 8601）",
            "example": "2026-08-04T10:00:00.000Z"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "alice",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "example": "Str0ng@Pass",
            "description": "密码"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "ValidateDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOi...",
            "description": "待校验的 Access Token"
          }
        },
        "required": [
          "token"
        ]
      },
...
      "CreateEnterpriseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "企业名称",
            "example": "腾讯科技有限公司"
          },
          "slug": {
            "type": "string",
            "description": "URL 友好标识（字母、数字、短横线），如 tencent",
            "example": "tencent"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateEnterpriseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "企业名称",
            "example": "腾讯科技有限公司"
          },
          "slug": {
            "type": "string",
            "description": "URL 友好标识",
            "example": "tencent"
          },
          "status": {
            "type": "string",
            "description": "企业状态：active / disabled",
            "example": "disabled"
          }
        }
      },
...
    }
  }
}